components were added
