+++ src/components/Todo/ActualTodo.tsx
@@ -1,0 +1,53 @@
+import classNames from 'classnames';
+import React from 'react';
+import { Todo } from '../../types/Todo';
+
+interface Props {
+  todo: Todo,
+}
+
+export const ActualTodo: React.FC<Props> = ({ todo }) => {
+  const { completed, title } = todo;
+
+  return (
+    <div
+      data-cy="Todo"
+      className={classNames(
+        'todo',
+        { completed: completed === true },
+      )}
+    >
+      <label className="todo__status-label">
+        <input
+          data-cy="TodoStatus"
+          type="checkbox"
+          className="todo__status"
+          defaultChecked
+        />
+      </label>
+
+      <span data-cy="TodoTitle" className="todo__title">
+        {title}
+      </span>
+
+      <button
+        type="button"
+        className="todo__remove"
+        data-cy="TodoDeleteButton"
+      >
+        ×
+      </button>
+
+      <div
+        data-cy="TodoLoader"
+        className="modal overlay"
+      >
+        <div className="
+          modal-background
+          has-background-white-ter"
+        />
+        <div className="loader" />
+      </div>
+    </div>
+  );
+};
